Fixed build for linux/ppc64le
